Add documentation for serialisation
This documentation covers Pony's serialisation system for pure-Pony
objects, as well as custom serialisation. It is meant to be used by
people who are interested in serialising and deserialising objects
between instances of the same executable by providing an overview of
how to use serialisation as well as working examples.